+++ src/app/configurator/shared/disabled-steps.type.ts
@@ -1,0 +1,5 @@
+export type DisabledSteps = {
+  1: boolean;
+  2: boolean;
+  3: boolean;
+};
