+++ src/app/configurator/shared/image.type.ts
@@ -1,0 +1,4 @@
+export type Image = {
+  src: string;
+  alt: string;
+};
