+++ src/app/configurator/shared/options.type.ts
@@ -1,0 +1,3 @@
+import { OptionsApiResponse } from '../../shared/services/options-api-response.type';
+
+export type Options = OptionsApiResponse;
